+++ examples/resources/trocco_job_definition/filter_column.tf
@@ -3,7 +3,6 @@ resource "trocco_job_definition" "filter_column_example" {
     {
       default                      = ""
       format                       = "%Y"
-      json_expand_columns          = []
       json_expand_enabled          = false
       json_expand_keep_base_column = false
       name                         = "id"
@@ -12,7 +11,6 @@ resource "trocco_job_definition" "filter_column_example" {
     },
     {
       default                      = "default value"
-      json_expand_columns          = []
       json_expand_enabled          = false
       json_expand_keep_base_column = false
       name                         = "name"
@@ -21,7 +19,6 @@ resource "trocco_job_definition" "filter_column_example" {
     },
     {
       default                      = ""
-      json_expand_columns          = []
       json_expand_enabled          = false
       json_expand_keep_base_column = false
       name                         = "timestamp"