+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.8.0
+FEATURES:
+- Added `salesforce` and `google_spreadsheets` type for `trocco_connection` resource.
+- Added `salesforce` input/output in `trocco_job_definition` resource.
+
+CHORE:
+- Remove `auto_create_table` option from `bigquery` output for `trocco_job_definition` resource.
+- Update documentation & examples
+
 ## 0.7.0
 FEATURES:
 - Added `snowflake` input/output in `trocco_job_definition` resource.